FLUME-3080. Close failure in HDFS Sink might cause data loss
If the HDFS Sink tries to close a file but it fails (e.g. due to timeout) the last block might
not end up in COMPLETE state. In this case block recovery should happen but as the lease is
still held by Flume the NameNode will start the recovery process only after the hard limit of
1 hour expires.

This change adds an explicit recoverLease() call in case of close failure.

This closes #127

Reviewers: Hari Shreedharan

(Denes Arvay via Bessenyei Balázs Donát)

Change-Id: I07f4b440538b9827c6a7b259a5b23d6f460bda87